Move views to own folder + use update vs updated in panel view
