--- src/panels/lovelace/hui-view-editable.ts
@@ -1,3 +1,0 @@
-// hui-view dependencies for when in edit mode.
-import "./components/hui-card-options";
-import "../../components/ha-fab";
