+++ src/panels/lovelace/views/hui-panel-view.ts
@@ -5,12 +5,12 @@ import {
   UpdatingElement,
 } from "lit-element";
 
-import applyThemesOnElement from "../../common/dom/apply_themes_on_element";
+import applyThemesOnElement from "../../../common/dom/apply_themes_on_element";
 
-import { HomeAssistant } from "../../types";
-import { LovelaceCard } from "./types";
-import { createCardElement } from "./common/create-card-element";
-import { LovelaceViewConfig } from "../../data/lovelace";
+import { HomeAssistant } from "../../../types";
+import { LovelaceCard } from "../types";
+import { createCardElement } from "../common/create-card-element";
+import { LovelaceViewConfig } from "../../../data/lovelace";
 
 @customElement("hui-panel-view")
 export class HUIPanelView extends UpdatingElement {
