+++ src/panels/lovelace/views/hui-view-editable.ts
@@ -1,0 +1,3 @@
+// hui-view dependencies for when in edit mode.
+import "../components/hui-card-options";
+import "../../../components/ha-fab";
